Add CeleritasG4 header and Celeritas::G4 target
