+++ example/geant4/add-celer.diff
